Add constant paths to new pkg/utils
This commit defines constants that will be referenced throughout the code in a single helper module, `pkg/utils`.

In order to ensure that we can dry-run the CCG Plugin Installer and CCG Plugin Account Provider on a non-Windows host, we use the local paths `dist/rancher-plugin-gmsa/ccg-plugin-installer` and `dist\rancher-plugin-gmsa\gmsa-account-provider` as the base directories when the host is non-Windows.

On a Windows host, we continue to use the host paths `C:\Program Files\RanchergMSACredentialProvider` and `C:\var\lib\rancher\gmsa`, respectively.